backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.24
python-dotenv==1.2.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
email-validator==2.3.0
numpy==1.26.4
pandas==2.2.2
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
pillow==10.4.0
reportlab==4.2.2
httpx==0.28.1
kagglehub==1.0.1
ultralytics>=8.0.0
